requirements.txt
sentence-transformers==3.3.1
chromadb==0.6.3
openai==1.68.2
python-dotenv==1.0.1
jupyter==1.1.1
notebook==7.3.3
numpy==2.2.4
scikit-learn==1.6.1
matplotlib==3.10.1
